Fix parsing issues
Fixes parsing issues, now both types of questions can be displayed.

+++ client/src/main/java/client/Session.java
@@ -9,7 +9,7 @@ public class Session {
 
     private static String nickname;
     private static String serverAddr;
-    private static int questionNum;
+    private static long questionNum;
 
     public static QuizzQuestionServerParsed emptyQ = new QuizzQuestionServerParsed(new QuizzQuestion("0",new Activity("0","0","0",Long.valueOf(0),"0"),new Activity("0","0","0",Long.valueOf(0),"0"),new Activity("0","0","0",Long.valueOf(0),"0")),-1,-1);
 
@@ -22,7 +22,7 @@ public class Session {
         return Session.serverAddr;
     }
 
-    public static int getQuestionNum() {
+    public static long getQuestionNum() {
         return Session.questionNum;
     }
 
@@ -34,7 +34,7 @@ public class Session {
         Session.serverAddr = serverAddr;
     }
 
-    public static void setQuestionNum(int questionNum) {
+    public static void setQuestionNum(long questionNum) {
         Session.questionNum = questionNum;
     }
 }
